2005-08-03  Jb Evain  <jbevain@gmail.com>
	* CodeGen/templates/OpCodes.cs
	  Mono.Cecil.Cil/OpCodes.cs
	  Mono.Cecil.Cil/OpCode.cs:
		Revert modifications to reduce code. OpCodes::.ctors were
		not called.

	* Mono.Cecil.Implem/ReflectionReader.cs:
		Make sure parameters are read even if there is no
		corresponding Param row. Thanks Cesar for the hint.

	* Mono.Cecil.Metadata/MetadataWriter.cs
	  Mono.Cecil.Metadata/UserStringsHeap.cs
	  Mono.Cecil.Metadata/GuidHeap.cs
	  Mono.Cecil.Implem/CodeReader.cs:
		On going work on the writing layer

svn path=/trunk/cecil/; revision=47979

+++ lib/Mono.Cecil.Implem/CodeReader.cs
@@ -143,7 +143,7 @@ namespace Mono.Cecil.Implem {
 					instr.Operand = br.ReadDouble ();
 					break;
 				case OperandType.InlineString :
-					instr.Operand = m_root.Streams.UserStringsHeap [(int) GetRid (br.ReadInt32 ())];
+					instr.Operand = m_root.Streams.UserStringsHeap [GetRid (br.ReadInt32 ())];
 					break;
 				case OperandType.InlineField :
 					int field = br.ReadInt32 ();
